requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
--trusted-host download.pytorch.org
Flask==2.0.2
jsonpickle==2.1.0
mediapipe==0.10.10
numpy==1.21.6
onnxruntime==1.11.0
opencv-python==4.5.5.62
pandas==1.3.5
python-dotenv==0.20.0
Pillow==9.1.0
requests==2.27.1
torch==1.10.1+cpu
torchvision==0.11.2+cpu
tqdm==4.62.3
protobuf==3.19.*
progressbar==2.5.*
Werkzeug==2.0.2
PyPDF2==3.0.1
PyMuPDF==1.19.6